backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.3.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
cors==1.0.1
fastapi-cors==0.0.6
